+++ CHANGELOG.md
@@ -26,6 +26,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to have too many `except` cases
 - Enforces to have an empty line after `coding` comment
 - Forbids to use too many `# noqa` comments
+- Forbids to use variables declared as unused
 
 ### Bugfixes